+++ modules/features/discover/src/main/java/au/com/shiftyjelly/pocketcasts/discover/view/DiscoverAdapter.kt
@@ -16,11 +16,13 @@ import androidx.recyclerview.widget.ListAdapter
 import androidx.recyclerview.widget.RecyclerView
 import androidx.recyclerview.widget.RecyclerView.OnScrollListener
 import au.com.shiftyjelly.pocketcasts.analytics.AnalyticsEvent
+import au.com.shiftyjelly.pocketcasts.analytics.AnalyticsEvent.DISCOVER_AD_CATEGORY_TAPPED
 import au.com.shiftyjelly.pocketcasts.analytics.AnalyticsTrackerWrapper
 import au.com.shiftyjelly.pocketcasts.analytics.FirebaseAnalyticsTracker
 import au.com.shiftyjelly.pocketcasts.discover.R
 import au.com.shiftyjelly.pocketcasts.discover.databinding.RowCarouselListBinding
 import au.com.shiftyjelly.pocketcasts.discover.databinding.RowCategoriesBinding
+import au.com.shiftyjelly.pocketcasts.discover.databinding.RowCategoryAdBinding
 import au.com.shiftyjelly.pocketcasts.discover.databinding.RowCategoryPillsBinding
 import au.com.shiftyjelly.pocketcasts.discover.databinding.RowChangeRegionBinding
 import au.com.shiftyjelly.pocketcasts.discover.databinding.RowCollectionListBinding
@@ -34,6 +36,7 @@ import au.com.shiftyjelly.pocketcasts.discover.databinding.RowSinglePodcastBindi
 import au.com.shiftyjelly.pocketcasts.discover.extensions.updateSubscribeButtonIcon
 import au.com.shiftyjelly.pocketcasts.discover.util.AutoScrollHelper
 import au.com.shiftyjelly.pocketcasts.discover.util.ScrollingLinearLayoutManager
+import au.com.shiftyjelly.pocketcasts.discover.view.DiscoverFragment.Companion.CATEGORY_ID_KEY
 import au.com.shiftyjelly.pocketcasts.discover.view.DiscoverFragment.Companion.EPISODE_UUID_KEY
 import au.com.shiftyjelly.pocketcasts.discover.view.DiscoverFragment.Companion.LIST_ID_KEY
 import au.com.shiftyjelly.pocketcasts.discover.view.DiscoverFragment.Companion.PODCAST_UUID_KEY
@@ -102,6 +105,7 @@ internal data class MostPopularPodcastsByCategoryRow(val listId: String?, val ca
     }
 }
 internal data class RemainingPodcastsByCategoryRow(val listId: String?, val category: String?, val podcasts: List<DiscoverPodcast>)
+internal data class CategoryAdRow(val discoverRow: DiscoverRow)
 internal class DiscoverAdapter(
     val context: Context,
     val service: ListRepository,
@@ -369,6 +373,7 @@ internal class DiscoverAdapter(
         private lateinit var source: String
         private lateinit var region: String
         private lateinit var allCategories: CategoryPill
+        private var mostPopularCategoriesId: List<Int> = emptyList()
 
         private val adapter = CategoryPillListAdapter(
             onCategoryClick = { selectedCategory, onCategorySelectionSuccess ->
@@ -421,9 +426,11 @@ internal class DiscoverAdapter(
                 adapter.loadCategories(categoriesFilter)
             }
         }
+        fun setMostPopularCategoriesId(ids: List<Int>) {
+            this.mostPopularCategoriesId = ids
+        }
         private fun getMostPopularCategories(categories: List<CategoryPill>): List<CategoryPill> {
-            // True Crime, Comedy, Culture, History, Fiction, Technology
-            val mostPopularCategoriesId = setOf(19, 3, 13, 18, 17, 15)
+            if (mostPopularCategoriesId.isEmpty()) return categories
 
             return categories
                 .filter { it.discoverCategory.id in mostPopularCategoriesId }
@@ -458,6 +465,7 @@ internal class DiscoverAdapter(
 
     class ErrorViewHolder(val binding: RowErrorBinding) : RecyclerView.ViewHolder(binding.root)
     class ChangeRegionViewHolder(val binding: RowChangeRegionBinding) : RecyclerView.ViewHolder(binding.root)
+    class CategoryAdViewHolder(val binding: RowCategoryAdBinding) : NetworkLoadableViewHolder(binding.root)
     class SinglePodcastViewHolder(val binding: RowSinglePodcastBinding) : NetworkLoadableViewHolder(binding.root)
     class SingleEpisodeViewHolder(val binding: RowSingleEpisodeBinding) : NetworkLoadableViewHolder(binding.root)
     class CollectionListViewHolder(val binding: RowCollectionListBinding) : NetworkLoadableViewHolder(binding.root)
@@ -474,6 +482,7 @@ internal class DiscoverAdapter(
             R.layout.row_category_pills -> CategoryPillsViewHolder(RowCategoryPillsBinding.inflate(inflater, parent, false))
             R.layout.row_most_popular_podcasts -> MostPopularPodcastsViewHolder(RowMostPopularPodcastsBinding.inflate(inflater, parent, false))
             R.layout.row_remaining_podcasts_by_category -> RemainingPodcastsByCategoryViewHolder(RowRemainingPodcastsByCategoryBinding.inflate(inflater, parent, false))
+            R.layout.row_category_ad -> CategoryAdViewHolder(RowCategoryAdBinding.inflate(inflater, parent, false))
             R.layout.row_single_podcast -> SinglePodcastViewHolder(RowSinglePodcastBinding.inflate(inflater, parent, false))
             R.layout.row_single_episode -> SingleEpisodeViewHolder(RowSingleEpisodeBinding.inflate(inflater, parent, false))
             R.layout.row_collection_list -> CollectionListViewHolder(RowCollectionListBinding.inflate(inflater, parent, false))
@@ -490,28 +499,25 @@ internal class DiscoverAdapter(
 
         when (row) {
             is DiscoverRow -> {
-                when (row.type) {
-                    is ListType.PodcastList -> {
-                        return when (row.displayStyle) {
-                            is DisplayStyle.Carousel -> R.layout.row_carousel_list
-                            is DisplayStyle.LargeList -> R.layout.row_podcast_large_list
-                            is DisplayStyle.SmallList -> R.layout.row_podcast_small_list
-                            is DisplayStyle.SinglePodcast -> R.layout.row_single_podcast
-                            is DisplayStyle.CollectionList -> R.layout.row_collection_list
-                            else -> R.layout.row_error
-                        }
+                if (row.type is ListType.PodcastList) {
+                    return when (row.displayStyle) {
+                        is DisplayStyle.Carousel -> R.layout.row_carousel_list
+                        is DisplayStyle.LargeList -> R.layout.row_podcast_large_list
+                        is DisplayStyle.SmallList -> R.layout.row_podcast_small_list
+                        is DisplayStyle.SinglePodcast -> R.layout.row_single_podcast
+                        is DisplayStyle.CollectionList -> R.layout.row_collection_list
+                        else -> R.layout.row_error
                     }
-                    is ListType.EpisodeList -> {
-                        return when (row.displayStyle) {
-                            is DisplayStyle.SingleEpisode -> R.layout.row_single_episode
-                            is DisplayStyle.CollectionList -> R.layout.row_collection_list
-                            else -> R.layout.row_error
-                        }
+                } else if (row.type is ListType.EpisodeList) {
+                    return when (row.displayStyle) {
+                        is DisplayStyle.SingleEpisode -> R.layout.row_single_episode
+                        is DisplayStyle.CollectionList -> R.layout.row_collection_list
+                        else -> R.layout.row_error
                     }
-                    is ListType.Categories -> {
-                        return if (FeatureFlag.isEnabled(Feature.CATEGORIES_REDESIGN)) R.layout.row_category_pills else R.layout.row_categories
-                    }
-                    else -> {}
+                } else if (row.type is ListType.Categories && row.displayStyle is DisplayStyle.Pills) {
+                    return R.layout.row_category_pills
+                } else if (row.type is ListType.Categories && row.displayStyle is DisplayStyle.Category) {
+                    return R.layout.row_categories
                 }
             }
             is ChangeRegionRow -> {
@@ -524,6 +530,10 @@ internal class DiscoverAdapter(
 
             is RemainingPodcastsByCategoryRow -> {
                 return R.layout.row_remaining_podcasts_by_category
+            }
+
+            is CategoryAdRow -> {
+                return R.layout.row_category_ad
             }
         }
 
@@ -619,6 +629,7 @@ internal class DiscoverAdapter(
                         loadCategories(row.source),
                         onNext = { categories ->
                             val context = holder.itemView.context
+                            row.mostPopularCategoriesId?.let { holder.setMostPopularCategoriesId(it) }
                             holder.submitCategories(categories, row.source, context, region = row.regionCode)
                         },
                     )
@@ -814,6 +825,41 @@ internal class DiscoverAdapter(
         } else if (row is RemainingPodcastsByCategoryRow) {
             val remainingPodcastHolder = holder as RemainingPodcastsByCategoryViewHolder
             remainingPodcastHolder.adapter.replaceList(row.podcasts)
+        } else if (row is CategoryAdRow) {
+            val adHolder = holder as CategoryAdViewHolder
+
+            (holder as NetworkLoadableViewHolder).loadFlowable(
+                loadPodcastList(row.discoverRow.source),
+                onNext = {
+                    val podcast = it.podcasts.firstOrNull() ?: return@loadFlowable
+                    val context = adHolder.itemView.context
+                    val podcastTitle = podcast.title
+
+                    adHolder.binding.lblTitle.text = podcastTitle
+                    adHolder.binding.lblBody.text = it.description
+
+                    imageRequestFactory.createForPodcast(podcast.uuid).loadInto(adHolder.binding.imgPodcast)
+                    adHolder.itemView.setOnClickListener {
+                        row.discoverRow.categoryId?.let { categoryId ->
+                            analyticsTracker.track(DISCOVER_AD_CATEGORY_TAPPED, mapOf(CATEGORY_ID_KEY to categoryId))
+                        }
+                        listener.onPodcastClicked(podcast, row.discoverRow.listUuid)
+                    }
+
+                    val lblSponsored = adHolder.binding.lblSponsored
+                    if (row.discoverRow.sponsored) {
+                        lblSponsored.setTextColor(context.getThemeColor(UR.attr.primary_text_02))
+                        lblSponsored.text = context.getString(LR.string.discover_row_sponsored)
+                    } else {
+                        lblSponsored.setTextColor(context.getThemeColor(UR.attr.support_02))
+                        lblSponsored.text = context.getString(LR.string.discover_row_fresh_pick)
+                    }
+
+                    val textSize = if ((podcastTitle ?: "").length < 15) 18f else 15f
+                    adHolder.binding.lblTitle.setTextSize(TypedValue.COMPLEX_UNIT_SP, textSize)
+                    onRestoreInstanceState(adHolder)
+                },
+            )
         }
     }