+++ modules/features/widgets/src/main/kotlin/au/com/shiftyjelly/pocketcasts/widget/SmallPlayerWidgetReceiver.kt
@@ -1,0 +1,6 @@
+package au.com.shiftyjelly.pocketcasts.widget
+
+internal class SmallPlayerWidgetReceiver : PlayerWidgetReceiver() {
+    override val glanceAppWidget = SmallPlayerWidget()
+    override val widgetTypeAnalyticsValue = "player_small"
+}