+++ modules/services/ui/src/main/java/au/com/shiftyjelly/pocketcasts/ui/helper/FragmentHostListener.kt
@@ -4,6 +4,7 @@ import android.view.View
 import androidx.fragment.app.Fragment
 import au.com.shiftyjelly.pocketcasts.models.type.EpisodeViewSource
 import com.google.android.material.bottomsheet.BottomSheetBehavior
+import kotlin.time.Duration
 
 interface FragmentHostListener {
     fun addFragment(fragment: Fragment, onTop: Boolean = false)
@@ -25,7 +26,13 @@ interface FragmentHostListener {
     fun getPlayerBottomSheetState(): Int
     fun addPlayerBottomSheetCallback(callback: BottomSheetBehavior.BottomSheetCallback)
     fun removePlayerBottomSheetCallback(callback: BottomSheetBehavior.BottomSheetCallback)
-    fun openEpisodeDialog(episodeUuid: String?, source: EpisodeViewSource, podcastUuid: String?, forceDark: Boolean)
+    fun openEpisodeDialog(
+        episodeUuid: String?,
+        source: EpisodeViewSource,
+        podcastUuid: String?,
+        forceDark: Boolean,
+        timestamp: Duration? = null,
+    )
     fun lockPlayerBottomSheet(locked: Boolean)
     fun updateSystemColors()
     fun overrideNextRefreshTimer()